+++ Classes/Scene/BattleScene/b2Sperator.h
@@ -21,7 +21,7 @@ class b2Separator
 public:
     b2Separator(){}
     
-    void separator(b2Body* body, b2FixtureDef* fixtureDef, std::vector<b2Vec2>* verticesVec, int scale=30);
+    void separator(b2Body* body, b2FixtureDef* fixtureDef, std::vector<b2Vec2>* verticesVec, float scale=30.f);
     int validate(const std::vector<b2Vec2>& verticesVec);
 
 private:
